First attempt at changing cursor while dragging notes.
This doesn't seem to work, and I'm not sure why.  Viewing the element
tree in the chrome dev tools shows that the beingdragged class is
getting added to the note while dragging, and the css for the
beingdragged class appears correct.  Perhaps there's something that
prohibits custom cursors on textareas?