Firefox does the flashing border thing when scrolling notes
It fades to the yellow border and then goes back, I think the only
difference, is that you can have multiple colored notes, but only for a
half a second.
It might be a good idea to have it clear the note that was previously
selected, but I wasn't sure how to do that.